COM-779 - move component to planning
